pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tls-dataset"
version = "0.1.0"
description = "Privacy-preserving encrypted traffic analytics platform for TLS 1.3 and QUIC."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Proprietary" }
authors = [
  { name = "Vandit Shah" }
]
dependencies = [
  "boto3>=1.35,<2",
  "beautifulsoup4>=4.12,<5",
  "fastapi>=0.115,<1",
  "matplotlib>=3.9,<4",
  "nfstream>=6.5,<7",
  "numpy>=2.1,<3",
  "pandas>=2.2,<3",
  "psycopg[binary]>=3.2,<4",
  "python-multipart>=0.0.18,<1",
  "PyYAML>=6.0,<7",
  "redis>=5.2,<6",
  "requests>=2.32,<3",
  "rq>=1.16,<2",
  "scapy>=2.5,<3",
  "scikit-learn>=1.5,<2",
  "SQLAlchemy>=2.0,<3",
  "uvicorn>=0.32,<1",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1",
  "mypy>=1.11,<2",
  "pytest>=8.3,<9",
  "ruff>=0.8,<1",
]

[project.scripts]
tls-dataset = "tls_dataset.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
packages = ["tls_dataset"]
warn_unused_configs = true
disallow_untyped_defs = true
